requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
databases
python-dotenv
asyncpg